src/api-service/requirements.txt
fastapi
uvicorn
google-cloud-storage
python-dotenv
flask
flask-bcrypt
flask-jwt-extended
flask-cors
passlib[bcrypt]
python-jose
python-multipart